# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: PortalDots API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080/v1
paths:
...
  /staff/pages/{pageID}/pin:
    parameters:
      - in: path
        name: pageID
        required: true
        schema:
          type: string
    patch:
      operationId: patchStaffPagePin
      summary: Update global notice pin status from staff mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isPinned:
                  type: boolean
      responses:
        "200":
          description: Updated page pin status
          content:
            application/json:
              schema:
                type: object
                required: [id, title, notes, createdAt, updatedAt, isPinned, isPublic, viewableTags, documentIds, documents]
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  notes:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
                  isPinned:
                    type: boolean
                  isPublic:
                    type: boolean
                  viewableTags:
                    type: array
                    items:
                      type: string
                  documentIds:
                    type: array
                    items:
                      type: string
                  documents:
                    type: array
                    items:
                      type: object
                      required: [id, name, description, isImportant, extension, sizeBytes, updatedAt, downloadUrl]
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                        isImportant:
                          type: boolean
                        extension:
                          type: string
                        sizeBytes:
                          type: integer
                          format: int64
                        updatedAt:
                          type: string
                          format: date-time
                        downloadUrl:
                          type: string
        "401":
          description: Unauthenticated
        "403":
          description: Staff verification is required
        "404":
          description: Page not found
        "409":
          description: Current circle is not selected
...